+++ include/Simulation.h
@@ -23,8 +23,7 @@ class Simulation {
         string file_suffix;
     
     public:
-        Simulation(int time_steps, int x_bound, int y_bound, int num_agents, int num_preadtors, int num_obstacles, 
-                   int num_energies, bool seek_energy, bool output_csv=true, bool use_advanced_agents=false);
+        Simulation();
         void output_csv_row(vector <string> outputs);
         void initialise_csv(BasicEnvironment environment, vector <BasicAgent*> agent_ptrs);
         void run_simulation(int time_delay=0);